Creating a Header folder in components. Moving the header.tsx and header.css files into it. Changing the route in app.tsx.
