+++ client/src/App.tsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import Header from "./components/Header";
+import Header from "./components/Header/Header";
 import DetailRecipePage from "./pages/DetailRecipePage/DetailRecipePage";
 
 function App() {
